Apply Black formatting

+++ example_affine.py
@@ -20,9 +20,9 @@ import numpy as np
 
 # Input image retrieved from:
 # https://raw.githubusercontent.com/skaae/transformer_network/master/cat.jpg
-im = imageio.imread('data/cat.jpg')
-im = im / 255.
-im = im.astype('float32')
+im = imageio.imread("data/cat.jpg")
+im = im / 255.0
+im = im.astype("float32")
 
 # input batch
 batch_size = 4
@@ -30,27 +30,28 @@ batch = np.expand_dims(im, axis=0)
 batch = np.tile(batch, [batch_size, 1, 1, 1])
 
 # Let the output size of the affine transformer be quarter of the image size.
-outsize = (int(im.shape[0]/4), int(im.shape[1]/4))
+outsize = (int(im.shape[0] / 4), int(im.shape[1] / 4))
 
 # Affine Transformation Layer
 stl = AffineTransformer(outsize)
 stlm = AffineTransformer(outsize, masked=True)
 
 # Identity transformation parameters
-initial = np.array([1.0, 0.0, 0.0,
-                    0.0, 1.0, 0.0]).astype('float32')
+initial = np.array([1.0, 0.0, 0.0, 0.0, 1.0, 0.0]).astype("float32")
 initial = np.reshape(initial, [1, stl.param_dim])
+
 
 def main(x):
     # Random jitter of the identity parameters
-    theta = initial + 0.1*tf.random_normal([batch_size, stl.param_dim])
+    theta = initial + 0.1 * tf.random_normal([batch_size, stl.param_dim])
     result = stl.transform(x, theta)
     resultm = stlm.transform(x, theta)
     return result, resultm
+
 
 result_, resultm_ = main(batch)
 
 # save our result
 for i in range(result_.shape[0]):
-  imageio.imsave('affine' + str(i) + '.png', result_[i])
-  imageio.imsave('affine_masked' + str(i) + '.png', resultm_[i])
+    imageio.imsave("affine" + str(i) + ".png", result_[i])
+    imageio.imsave("affine_masked" + str(i) + ".png", resultm_[i])
