+++ example_elastic.py
@@ -19,9 +19,9 @@ import numpy as np
 
 # Input image retrieved from:
 # https://raw.githubusercontent.com/skaae/transformer_network/master/cat.jpg
-im = imageio.imread('data/cat.jpg')
-im = im / 255.
-im = im.astype('float32')
+im = imageio.imread("data/cat.jpg")
+im = im / 255.0
+im = im.astype("float32")
 
 # input batch
 batch_size = 4
@@ -29,27 +29,27 @@ batch = np.expand_dims(im, axis=0)
 batch = np.tile(batch, [batch_size, 1, 1, 1])
 
 # Let the output size of the elastic transformer be quarter of the image size.
-outsize = (int(im.shape[0]/4), int(im.shape[1]/4))
+outsize = (int(im.shape[0] / 4), int(im.shape[1] / 4))
 
 # Elastic Transformation Layer
 stl = ElasticTransformer(outsize)
-stlc = ElasticTransformer(outsize, interp_method='bicubic')
-stlcm = ElasticTransformer(outsize, interp_method='bicubic',
-                           masked=True, cval=0.5)
+stlc = ElasticTransformer(outsize, interp_method="bicubic")
+stlcm = ElasticTransformer(outsize, interp_method="bicubic", masked=True, cval=0.5)
 
 # Run session
 def main(x):
     # Random jitter of identity parameters
-    theta = 0.1*tf.random_normal([batch_size, stl.param_dim])
+    theta = 0.1 * tf.random_normal([batch_size, stl.param_dim])
     result = stl.transform(x, theta)
     resultc = stlc.transform(x, theta)
     resultcm = stlcm.transform(x, theta)
     return result, resultc, resultcm
 
+
 result_, resultc_, resultcm_ = main(batch)
 
 # save our result
 for i in range(result_.shape[0]):
-  imageio.imsave('elastic' + str(i) + '.png', result_[i])
-  imageio.imsave('elastic' + str(i) + '_bicubic.png', resultc_[i])
-  imageio.imsave('elastic' + str(i) + '_bicubic_masked_grey.png', resultcm_[i])
+    imageio.imsave("elastic" + str(i) + ".png", result_[i])
+    imageio.imsave("elastic" + str(i) + "_bicubic.png", resultc_[i])
+    imageio.imsave("elastic" + str(i) + "_bicubic_masked_grey.png", resultcm_[i])
